Clear BFD poll bit if final bit is being set
When a session transitions to UP state, it is necessary to update
the MinTxInterval, and this is initiated by setting bfd->poll so
that when the next packet is sent, the poll bit will be set.
However, if we then receive a packet with the poll bit set, we
were also setting both the poll and final bit in the reply packet
which is invalid.

If the final bit is set in a packet, we can send updated parameters
without setting the poll bit, so simply clear the poll bit if the
final bit is set.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>