When BFD not in UP state, idle_min_tx_intv should be sent
The Required Min Rx Interval must not be less than 1,000,000 micro
seconds when not in UP state, and this needs to be set accordingly
in the BFD packets being sent.

If local_idle_tx_intv != local_min_tx_intv, then when we transition
to UP state a poll/final sequence will also need to be initialed
to change the RequiredMinRxInterval.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>