Ensure BFD slow (idle) tx timer used when not in UP state
If a packet was received when not in UP state, the local_tx_interval
was being set to local_min_tx_intv rather than local_idle_tx_intv
causing fast transmission of packets when not in up state.

The commit now checks the state to ensure that the correct
idle_tx_intv value is used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>