Allow BFD parameters to be any legal value
This commit changes the ranges of the configurable BFD parameters
to be anything supported by the protocol, subject to being configurable
in milli-seconds, whereas previously the limits were artifically
constrained.

If any values configured are outside the previous constraints, then a
warning is logged that the configured values may not be sensible.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>