Improve and fix use of getopt_long()
We musn't use a long option val of 1, since getopt_long() can return
that value.

getopt_long() also returns longindex == 0 when there is no matching
long option, and there needs to be careful checking if there is an
error to work out whether a long or short option was used, which is
needed for meaningful error messages.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>