Fix some SNMP keepalived checker integer types and default values
Some virtual server and real server values were being sent to SNMP
with a signed type whereas the value is unsigned, so set the type
field correctly.

Some virtual server and real server values that apply to checkers
are set to nonsense default values in order to determine if a
value has been specified. Handle these values when reporting them
to SNMP replying with 0 rather than a nonsense value.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>