Improve the log message when a master receives higher priority advert
The log message reported in issue #754
"VRRP_Instance(VI_1) Received advert with higher priority 253, ours 253"
is somewhat misleading since 253 == 253.

This commit improves the log message in this case be reporting that
the sender's IP address is higher and the priority is equal. It also
states the it was a master receiving the advert.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>