Improve efficiency of using netlink
At startup when acquiring interface and address details, use the
nl_cmd netlink socket rather than opening a new netlink socket
for each interface created.

Also, the nl_cmd netlink socket is always used synchronously so don't
set SOCK_NONBLOCK, and this avoids having to set the socket to
blocking whenever we want to read the response to a command.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>