Allow more flexibility of layout of { and } in config files
keepalived was a bit fussy about where '{'s and '}'s (braces) could
be placed in terms of after the keyword, or on a line on their own.
It certainly was not possible to have multiple braces on one line.

This commit now provides complete flexibility of where braces are, so
long as they occur in the correct order.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>