By default, tell netlink not to report ENOBUFS
There appears to be a kernel bug that manifests itself when we have a large number
of VMAC interfaces to add (i.e. 200 or more). After approx 200 interfaces have been
added the kernel will return ENOBUFS on the nl_kernel socket, and then repeat the
first 30 or so RTM_NEWLINK messages, omitting the first one. Then, at the end of
creating all the interfaces, i.e. after a slight delay with no new messages,
we get another ENOBUFS and all the RTM_NEWLINK messages from the time of the
first ENOBUFS message repeated.

This problem feels as though a circular buffer is wrapping around, and causes
all the old messages in the buffer to be resent, but the first one is omitted.
Note that it is only the interfaces that keepalived creates that are resent,
not interfaces that already existed on the system before keepalived starts.

We can also get ENOBUFS on the nl_cmd socket if the NLM_F_ECHO flag is set as well as
the NLM_F_ACK flag when a command is sent on the nl_cmd socket.

It appears that this must be a kernel bug, since when it happens on interface creation,
if we are also running `ip monitor link addr route`, i.e. the same as the nl_kernel
socket, then precisely the same messages are repeated (provided we have set the buffer
IF_DEFAULT_BUFSIZE (the rcvbuf size) to be 1024 * 1024 to match what ip monitor does).

Further investigations are needed to identify the exact cause of the problem, and
if the kernel is patched to resolve the problem, then we can remove this work-around
iwhen running on kernels that have the fix.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>