fixes #761: INSTALL: note linux distro package that provides 'sphinx_rtd_theme'
This changeset updates the individual sections for each of the four Linux
distro-lineages that currently have a section in the `INSTALL` file
(RedHat-based, Debian-based, Alpine Linux, and archlinux). The changes just
note the distro-specific package names for the tools needed to build the
HTML-based documentation (basically, some variant of the `sphinx-build` tool
and `sphinx_rtd_theme`) and (except for Alpine Linux[0]) the LaTeX-based PDF
documentation.

The changes were tested using `vagrant` and the vagrant "boxes" published by
respective distros (and/or their trusted volunteers) for the latest stable
version of each distro.

   * RedHat-based
     - centos/7              (virtualbox, 1801.02)
     - fedora/27-cloud-base  (virtualbox, 20171105)

   * Debian-based
     - debian/stretch64      (virtualbox, 9.3.0)
     - ubuntu/artful64       (virtualbox, 20180126.0.0)

   * Alpine Linux
     - alpine/alpine64       (virtualbox, 3.6.0)

   * archlinux
     - archlinux/archlinux   (virtualbox, 2018.01.07)

[0] LaTeX/PDF documentation build on Alpine Linux currently requires a
    roll-your-own approach, as texlive is not yet a supported package for that
    distro. See also:

        https://bugs.alpinelinux.org/issues/3920
        https://bugs.alpinelinux.org/issues/4969