Clear /proc/sys/net/ipv6/conf/IF/disable_ipv6 when create VMACs
An issue was identified where keepalived was reporting permission
denied when attempting to add an IPv6 address to a VMAC interface.
It turned out that this was because
/proc/sys/net/ipv6/conf/default/disable_ipv6
was set to 1, causing IPv6 to be disables on all interfaces that
keepalived created.

This commit clears disable_ipv6 on any VMAC interfaces that
keepalived creates if the vrrp instance is using IPv6.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>