Stop backup taking over as master while master reloads
If a reload was initiated just before an advert, and since it took
one advert interval after a reload before an advert was sent, if the
reload itself took more than one advert interval, the backup could
time out and take over as master.

This commit makes keepalived send adverts for all instances that are
master immediately before a reload, and also sends adverts immediately
after a reload, thereby trippling the time available for the reload
to complete.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>