Don't set ipvs sync daemon to master state before becoming master
If a vrrp instance which was the one specified for the ipvs sync
daemon was configured with initial state master, the sync daemon
was being set to master mode before the vrrp instance transitioned
to master mode. This caused an error message when the vrrp instance
transitioned to master and attempted to make the sync daemon go from
backup to master mode.

This commit stops setting the sync daemon to master mode at initialisation
time, and it is set to master mode when the vrrp instance transitions
to master.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>