Remove duplication of code for writing config to file and syslog
vrrp_print.c very nearly duplicated the code in vrrp_data.c for
writing the configuration to /tmp/keepalived.data and to the syslog
when using the '-d' command line option. This commit merges the code
so that the code in vrrp_data.c can write to either the file or to
syslog, and the code is removed from keepalived_print.c.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>