Use configure to check for SSL_set_tlsext_host_name()
Rather than checking for a specific version of the OpenSSL library
(and it would also need checking the version of the LibreSSL library)
let configure check for the presence of SSL_set_tlsext_host_name().

Also omit all code related to SNI of SSL_set_tlsext_host_name() is
not available.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>