Open and run the notify fifo and script if no other fifo
Due to the way the code was structured the notify_fifo for both
checker and vrrp messages wasn't run if neither the vrrp or checker
fifo wasn't configured.

Also, if all three fifos were configured, the general fifo script
was executed by both the vrrp and checker process, causing problems.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>