Make checker process handle RTM_NEWLINK messages with -a option
Even though the checker process doesn't subscribe to RTNLGRP_LINK
messages, it appears that older kernels (certainly 2.6.32) can
send RTM_NEWLINK (but not RTM_DELLINK) messages. This occurs
when the link is set to up state.

Only the VRRP process is interested in link messages, and so the
checker process doesn't do the necessary initialisation to be able
to handle RTM_NEWLINK messages.

This commit makes the checker process simply discard RTM_NEWLINK
and RTM_DELLINK messages, rather than assuming that if it receives
an RTM_NEWLINK message it must be the VRRP process.

This problem was reported in issue #848 since the checker process
was segfaulting when a new interface was added when the -a command
line option was specified.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>