Fix building on Fedora 28
net-snmp-config output can include compiler and linker flags that
refer to spec files that were used to build net-snmp but may not
exist on the system building keepalived. That would cause the build
done by configure to test for net-snmp support to fail; in particular
on a Fedora 28 system that doesn't have the redhat-rpm-config package
installed.

This commit checks that any spec files in the compiler and linker
flags returned by net-snmp-config exist on the system building
keepalived, and if not it removes the reference(s) to the spec file(s).

See https://bugzilla.redhat.com/show_bug.cgi?id=1544527 for further
details.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>