Implement two phase shutdown
When shutting down keepalived shutdown scripts potentially need to
be called. We need to allow the scripts to complete, and also ensure
that they have completed, so we need to have a two phase shutdown,
so that in phase 1 most processing can stop, running scripts can be
stopped, but also shutdown notify scripts can be invoked. The second
phase makes sure the shutdown scripts have completed before keepalived
exits.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>