Change the way packages are installed by Travis-CI
Travis-CI builds has started failing due to packages not being
installed. The following was listed on the Travis-CI status page:

Update - We have changed the build script to not default to running
apt-get update for the time being, unless we're able to auto-detect
that your build uses apt-get install. If your build requires it in
other ways, please do opt in via .travis.yml. The build log gives
instructions on how to opt in.
May 9, 15:23 UTC

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>