Make the signal value checking code only report each signal once
If incorrect signals are being received we don't want to flood the
logs with messages, so only report an invalid signal number if the
signal is outside the range of signals already received.

Also, this is essentially debug code, so wrap it in conditional
compilation so that it can be enabled when needed.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>