Add no_checker_emails onfig option
By default, if smtp_alert is set, an email is sent whenever a checker
a checker changes from failed to successful or vice versa. This commit
adds option no_checker_emails so that emails are not sent whenever a
checker changes state but only when a real server goes up or down.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>