Add warning if virtual and real servers are different address families
Although normally the virtual server and real servers must have the
same address family, if a real server is tunnelled, the address families
can be different. However, the kernel didn't support that until 3.18,
so add a check that the address families are the same if different
address families are not supported by the kernel.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>