Add no-track option for VIPs/eVIPs
This option allows VIPs/eVIPs to be configured, but the deletion
of those addresses will not cause the VRRP instance to transition
to backup state.

Merge request #860 gives a justification for why there are
circumstances when a vrrp instance should not transition to fault
state if a VIP is deleted.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>