Add tracking of virtual routes
If a virtual route is deleted, by default to vrrp instance will
now transition to backup mode, and if it transitions to master
again the route will be re-added. If an interface on which a
route is configured is down, then the instance will go to fault
state, since the route cannot be added.

This commit also adds a no-track option for routes, which means
that deletion of the route will not cause the vrrp instance to
transition to backup.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>