Set default preferences for ip rules if not specified
Since different vrrp instances can become master in different
orders, if preferences (priorities) are not specified for
ip rules, the order in which they are specified will be
indeterminate. In order to give some consistency, keepalived
will not allocate a default preference to each rule, and will
also warn that this is probably not going to work as intended.

The solution is to specify a preference for each rule.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>