Add tracking of static addresses, routes and rules
By default a static address, route or rule will now be reinstated
if it is deleted, unless the no_track option is specified.

In addition, if a track_group is specified for an address/route/rule
then if the address/route/rule cannot be reinstated (e.g. if the
specified interface is down or has been deleted), then the vrrp
instances specified in the track group will transition to fault state
until the interface comes back up.

This commit completes the monitoring and reinstatement of addresses
routes and rules and means that keepalived should now fully support
hot-swap devices.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>