Stop receive message queues not being read on send sockets
We shouldn't receive anything on vrrp send sockets since IP_MULTICAST_ALL is
cleared, and no multicast groups are subscribed to on the socket. However,
Debian Jessie with a 3.16.0 kernel, CentOS 7 with a 3.10.0 kernel,
and Fedora 16 with a 3.6.11 kernel all exhibit the problem of multicast
packets being queued on the send socket. Whether this was a kernel problem
that has been subsequently resolved, or a system default configuration
problem isn't yet known.

The workaround to the problem is to read on the send sockets, and to discard
any received data.

If anyone can provide more information about this issue it would be
very helpful.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>