Add options to set vrrp socket receive buffer sizes
Some systems have very large settings for net.core.rmem_default
allowing very large receive queues to build if the sockets aren't
read. keepalived doesn't need large buffers for receive queues,
so this commit allows options for setting the maximum buffer sizes
to be much smaller. It also adds the option of setting the receive
buffer size on the vrrp send sockets to be as small as possible,
since we shouldn't be receiving anything on those.

Following commit 6fb5980 - "Stop receive message queues not being
read on send sockets", this commit also adds the option not to
read the send sockets, which can be used where it is known that the
kernel will not queue unwanted multicasts to the send socket.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>