Add -t config-test option
Issue #389 has received increasing support to add a configuration
validation option. This commit adds the -t/--commit-test option
to report any detected configururation errors and exit.

Errors are logged to the system log by default, but use of the -g
and -G options can make the errors be logged to files.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>