Fix setting i/f params on a bridge underlying i/f of a VMAC
Issue #944 identified that when the underlying interface of a VMAC
interface was a bridge, keepalived was failing to set arp_ignore and
arp_filter in the underlying bridge interface. The problem appears to
lie in the libnl3 library. The description of the problem given in the
issue report was:
  Problem is that ifi_family is set to AF_BRIDGE, whereas it should be set
  to AF_UNSPEC. The kernel function that handles RTM_SETLINK messages for
  AF_BRIDGE doesn't know how to process the IFLA_AF_SPEC attribute.

This commit stops using libnl3 for setting/clearing arp_ignore and
arp_filter, and directly constructs the netlink messages in keepalived.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>