Remove cmd_str from notify_script_t
The cmd_str string (sort of) duplicated what was in the args array
of a notify_script_t, but was not always accurate. With the removal
of cmd_str, whenever it needs to be output, the string is now
generated from the args array, so accurately reflects what is
actually executed.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>