Don't run a sublevel close handler on a skipped configuration block
If a configuration block was skipped due to an error, the configuration
read won't be valid and may not even exist, so make sure the sublevel
end handler isn't run.

An example is if a virtual_server block is skipped, then the sublevel
end handler would have run against the previous (if any) virtual_server,
and if there hadn't been a previous virtual_server block it could
segfault.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>