Prepend WARNING to read_int/unsigned/double messages if not rejecting
read_int()/read_unsigned()/read_double() can output log messages if
the syntax is invalid but the configuration is being accepted, e.g.
parsing 12zyx will return 12 if _STRICT_CONFIG_ is not defined. In
this case we want to indicate that the entry is not valid, but we
are still processing it, so prepend the error message with WARNING.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>