Only add garp_delay_t's to interfaces used by VRRP instances
There is no point allocating a garp_delay_t to an interface that
isn't used by a VRRP instance, so this commit make keepalived only
allocate garp_delay_t scructures to the used interfaces.

In addition, when the configuration is dumped, list the interfaces
relevant to each garp_delay_t.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>