Fix a multiline parameter substitution having a replacement on 1st line
If a multiline parameter definition had a replaceable parameter or a
definition on the first non-blank line, it wasn't being handled. This commit
ensures that replaceable parameters/definitions on the first line are
handled correctly.

Example:

$RS= \
$RS_NUM=1 \
...

When $RS was used, $RS_NUM=1 would not be recognised as a new definition
and was reported as an unknown keyword.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>