Don't allow use_vmac on a macvlan (vmac) interface
Issue #998 reported problems when configuring a vrrp instance to
use a vmac when the underlying interface was itself a vmac. The
problem was that if a macvlan is configured on top of another
macvlan the kernel adds the new macvlan on the underlying interface
of the original macvlan, rather than on the original macvlan itself.

The relevant kernel code in drivers/net/macvlan.c is the following in
function macvlan_common_newlink():

 /* When creating macvlans or macvtaps on top of other macvlans - use
  * the real device as the lowerdev.
  */
 if (netif_is_macvlan(lowerdev))
         lowerdev = macvlan_dev_real_dev(lowerdev);

This commit now will not allow a vrrp instance to specify use_vmac
when the interface specified is a macvlan and will reject the
configuration. If dynamic interfaces are configured (i.e. interfaces
being deleted and recreated are supported) then when an interface
specified as the interface of a vrrp instance is created, if the
vrrp instance is configured to use a vmac, the created interface is
checked to see if it is a macvlan, and if so then it is ignored.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>