If receive lower priority advert, send GARP messages for sync group
A recent update to issue #542 identified that following recovery
from a split brain situation, GARP messages weren't being sent. It
transpired that, if a member of a sync group in master state received
a lower priority advert and vrrp_higher_prio_send_advert is set, a
further (lower priority) advert is sent, and the instance and all the
members of the sync group transition to backup (the other members of
the sync group don't send a further advert since they haven't received
a higher priority advert). This meant that the other members of the
sync group on the keepalived instance that remained master didn't
receive a lower priority advert, and so didn't send further GARP
messages.

This commit changes keepalived's behaviour, so that if a vrrp instance
is sending GARP messages due to receiving a lower priority advert
and it is a member of a sync group, keepalived will also send GARP
messages for any other member of the sync group that have
garp_lower_prio_rep set.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>