Fix effective_priority after reload if tracked interface down
Issue #1009 identified that after changing both the configured
priority of a vrrp instance and the weight of a tracked interface
the effective priority after a reload was wrong.

It was discovered that this was due to effective priority being
calculated solely from the old effective priority and the change in
configured priority, overriding earlier calculations that attempted
to calculate the effective priority taking into account the weights
as well.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>