Use an rbtree for finding vrrp instance for received advert
Previously the code search a list of pointers to vrrp instances and
looked for a matching fd and vrid. In order to optimise this, it was
implemented using an mlist whose index was a hash of the fd and vrid.

This commit changes the approach and uses an rbtree for each sock_t.
Since the sock_t that the advert was received on is known, the rbtree
search is only searching for a match on the vrid.

Not only is this more efficient, but it is simpler, uses standard code,
and reduces the code by over 60 lines.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>