Use an rbtree for finding vrrp instance for socket timeout
Previously the code search a list of pointers to vrrp instances and
looked for matching file descriptor and sands < time_now. In order to
optimise this, it was implemented using an mlist whose index was a hash
of the fd.

This commit changes the approach and uses a second rbtree for each sock_t.
Since the sock_t that the timeout occurred on is known, the rbtree
search is only searching for a match of the sands.

Not only is this more efficient, but it is simpler, uses standard code,
and reduces the code by over 220 lines.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>