Update rb_for_each_entry etc and rb_move to use rb_entry_safe
With the added definition of rb_entry_safe in the rbtree code
updated to Linux 4.18.10, the refinition of rb_entry was reverted
to the kernel definition. That meant that rb_for_each_entry,
rb_for_eacn_entry_safe and rb_move neded to be updated to use
rb_entry_safe rather than rb_entry.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>