Use cached rbtrees where the key is a timeval_t sands
When the key of an rbtree is a timeval_t sands keepalived will frequently
need to access the first node of the tree in order to calculate the next
timeout. This applies to the read, write, child and timer threads queues,
and also the vrrp queues on a sock_t.

The use of cached rbtrees for these is ideal since it gives direct access
to the first node of the queue.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>