Add thread_add_read_sands to avoid introducing timer errors
When using thread_add_read and the timeout was held as timeval_t,
it was converted to and offset from time_now, and then converted
back to a timeval_t, but time_now was updated, resulting in a
slightly different value being used as the timeout.

Using thread_add_read_sands() avoids the double conversion and
results in the timeout being more accurate.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>