Ensure all members of sync group have same preempt/nopreempt setting
If they don't match, then some instances will attempt to become master
and some will remain backup, and keepalived enters a tight loop.

This commit sets all member of a sync group to the same setting, which
may be nopreempt or preempt depending on whether any instance is an
address owner or wants to start in master state.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>