Fix segfault when receive netlink message for default route added
Issue #1025 identified a segfault when running keepalived in a
container with no addresses assigned to interfaces. It transpires
the problem was relating to adding a default route, when the kernel
sends an RT_NEWLINK message with no RTA_DST field, but the code
assumed that there always was an RTA_DST field. The code now
explicity checks if there is an RTA_DST field, and if not constructs
a default address to pass to the compare_addr() function.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>