Check for VRID conflicts when changeable interfaces are added
For example, a vrrp instance could be configured on a macvlan, and
that macvlan could be deleted and recreated with another base interface.
The VRIDs in this case need to be checked for duplicates against the
base interface, and so the VRID check needs to be done dynamically.

In order to allow VRID conflicts to produce config errors at startup,
by default keepalived assumes that there won't be interface movements
as described above, and will only handle it if the global_defs option
'dynamic_interfaces' is used along with the option 'allow_if_changes'.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>