Minor improvements re thread handling and cancellation
Issue #1037 identified that there was a problem when timer threads
were cancelled, since it could try removing the thread from the
timer queue, event if it had been moved to the ready queue.

This problem was fixed when implementing the scheduler-ng, since
thread types are now changed when moving threads from the waiting queues
to the ready queue, and the thread type is used to know what queue
the thread is on. There was one exception to this; a THREAD_CHILD
could be on either the child queue or the ready queue, although this
didn't cause a problem in practice since child threads are never
cancelled.

This commit adds a thread type THREAD_CHILD_TERMINATED, for use on the
ready queue, and THREAD_CHILD is used on the child queue.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>