Fix updating a timer thread's timeout
Issue #1042 identified that the BFD process could segfault. This
was tracked down to a timer thread which had already expired having
its timeout updated by timer_thread_update_timeout().

The sands timer should only be updated if the thread is on a waiting
queue, and not if it has already timed out or it is unused.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>