Improve mem-check diagnostics
When using an allocation list of over 50,000 entries, it was quite slow
searching thtough all the entries to find the matching memory allocation,
and to find free entries. This commit changes to using malloc() to create
entries, and a red-black tree to hold the entries. It also has a separate
list of free entries.

This commit also adds 4 more types of memory allocation error, and
improves the consistency of the entries in the log files.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>