Handle the underlying interface of a macvlan interface going up/down
The kernel doesn't send netlink messages for macvlans going up or
down when the underlying interface transitions (it doesn't even
update their status to say they are up/down), but the interfaces
don't work. We need to track the state of the underlying interfaces
and propagate that to the macvlan interfaces.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>