fix socket allocation with dynamic interfaces
When there are several vrrp instance binding different interfaces that
don't exist at startup, their ifindex is set to 0 in the sock. The
function already_exist_sock() that lookup for an existing socket will
always return the first sock because the ifindex is the same.

Later, when an interface appears, the fd will be created for one
instance, and all instances will wrongly use this fd to send the
advertisments.

Fix this by using the interface name instead of the ifindex as a key
for sock lookup.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>