Fix building with IP_MULTICAST_ALL in linux/in.h but not netinet/in.h
Issue #1054 identified that configure was checking the definition of
IP_MULTICAST_ALL in linux/in.h but including netinet/in.h, which also
has the definition, but only from glibc 2.17.

This commit creates a local definition (in lib/config.h) of IP_MULTICAST_ALL
if it is defined in linux/in.h but not in netinet/in.h. The reason for
this is that compiles using linux/in.h fail due to conflicting definitions.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>