Fix segfault while shutting down when SNMP activity occurs
Issue #1061 identified that keepalived could segfault when it
shut down. It appears that this was caused by data being received
on the file descriptors that the snmp agent requests keepalived
to monitor with epoll(). Since the read threads weren't being
processed during a shutdown, the first time an snmp fd was ready,
keepalived discarded the read thread. The second time that fd became
ready there was no thread to handle the fd, and, since the assert()
statement was not compiled in, non existant data was queued to the
thread ready queue.

This commit changes the assert() calls to continue, so that non existant
data is no longer queued to the thread ready queue.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>