Fix double free when global data smtp_helo_name copied from local_name
Issue #1071 identified a double free fault. It occurred when smtp_helo_name
was not set, in which case it was set to point to the same malloc'd memory
as local_name. At termination keepalived freed both local_name and
smtp_helo_name.

If keepalived needs to use local_name for smtp_helo_name it now malloc's
aadditional memory to copy the string into.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>