Add track_process for vrrp to monitor if another process is running
Configurations frequently include a track_script to check that a process
is running, often haproxy or nginx. Using any of pgrep, pkill, killall,
pidof, etc, has an overhead of reading all /proc/[1-9]*/status and/or
/proc/[1-9]*/cmdline files. In particular reading the cmdline files
has a significant overhead on a system that is swapping, since the
cmdline files provide access to part of the address space of each
process, which may need to be fetched from the swap space.

This commit reads the /proc/[1-9]*/stat and/or the /proc/[1-9]*/cmdline
files only when keepalived starts, and after that uses the process events
connector to track process creation and termination.

keepalived will ignore zombie processes, whereas pgrep etc include them.

A minimum number of instances of a process can be specified, and also a
delay so that if a process is restarted, it won't cause monitoring vrrp
instances to immediately transition to fault state but to wait the
configured time and it the monitored process starts again it
won't transition to fault state.

There are potential difficulties with the process event connector if a
large number of process events occur very rapidly, since there can be
a receive buffer overrun on the netlink socket. This code will detect
that happening, increase the receive buffer size, and reread the processes
from /proc.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>