Stop vrrp process entering infinite loop when track script times out
Issue #1093 identified that the vrrp process was entering an infinite
loop after a track script timed out. This was due to a child process
thread having an RB tree for PIDs as well as for the timeout, and if
a child process timed out, the thread wasn't being removed from the
PID RB tree. This commit now ensures it is removed.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>